requirements.txt
pydantic[email]~=1.10.2
motor~=3.0.0
fastapi~=0.85.1
pymongo~=4.3.2
SQLAlchemy~=1.4.42
python-decouple
neomodel
neo4j